# region: optimisee decompressCellblocks